Created household.py table
